chore(1-3336): rename files
