+++ frontend/src/component/admin/network/NetworkTrafficUsage/traffic-calculations.ts
